Optimize Input component
